refactor(maci-core): refactor directory structure for quick core component access
rearranged the 'core/ts' directory to spotlight the 'MaciState' and 'Poll' classes, improving the ease of identifying key functionalities. Shifted utility functions into a separate 'lib' subdirectory to streamline navigation and reduce search time for primary elements.
