+++ core/ts/index.ts
@@ -1,11 +1,21 @@
+export { MaciState } from "./MaciState";
+
 export {
   //genPerVOSpentVoiceCreditsCommitment,
   //genSpentVoiceCreditsCommitment,
-  MaciState,
   Poll,
+} from "./Poll";
+
+export {
+  MaxValues,
+  TreeDepths,
+  BatchSizes,
   genProcessVkSig,
   genTallyVkSig,
   genSubsidyVkSig,
-  MaxValues,
-  TreeDepths,
-} from "./MaciState";
+  packProcessMessageSmallVals,
+  unpackProcessMessageSmallVals,
+  packTallyVotesSmallVals,
+  unpackTallyVotesSmallVals,
+  packSubsidySmallVals,
+} from "./utils/utils";
